+++ Aikuma/src/org/lp20/aikuma/audio/record/Recorder.java
@@ -33,7 +33,7 @@ public class Recorder implements AudioHandler, MicrophoneListener, Sampler {
 	 * Creates a Recorder that uses an analyzer which tells the recorder to
 	 * always record regardless of input.
 	 */
-	public Recorder(File path, int sampleRate) throws MicException {
+	public Recorder(File path, long sampleRate) throws MicException {
 		this(path, sampleRate, new SimpleAnalyzer());
 	}
 
@@ -44,7 +44,7 @@ public class Recorder implements AudioHandler, MicrophoneListener, Sampler {
 	 * @param Pass in an analyzer which decides whether
 	 *        the recorder should record or ignore the input.
 	 */
-	public Recorder(File path, int sampleRate, Analyzer analyzer) throws MicException {
+	public Recorder(File path, long sampleRate, Analyzer analyzer) throws MicException {
 		this.analyzer = analyzer;
 		setUpMicrophone(sampleRate);
 		setUpFile();
@@ -75,7 +75,7 @@ public class Recorder implements AudioHandler, MicrophoneListener, Sampler {
 	}
 
 	/** Sets up the micrphone for recording */
-	private void setUpMicrophone(int sampleRate) throws MicException {
+	private void setUpMicrophone(long sampleRate) throws MicException {
 		this.microphone = new Microphone(sampleRate);
 	}
